Filter URLs Filter URLs in the provided database based on topic, country and language

Filter URLs Filter URLs in the provided database based on topic, country and language

filter_urls(
  topic = NULL,
  country = NULL,
  language = NULL,
  rss_table = package_rss
)

Arguments

topic
the topic of the feed see `show_topics()` for more info.

country
the country of origin of the feed using two capital letters, for example "US". See `show_countries()` for more info.

language
the language of the content of the feed using two lowercase letters, for example "en". See `show_languages()` for more info.

rss_table
a dataframe with urls and rss feeds in case you need to construct your own out of websites not in the included database. #Be sure to have the same format as the included data. See `R/package_rss.R` #for details.

Value

a dataframe filtered according to the given parameters

Examples

filter_urls(topic = "tech", country = "US", language = "en")

#>                clean_url language topic_unified main clean_country
#> 33       ycombinator.com       en          tech    1            US
#> 71               wsj.com       en          tech None            US
#> 115            wired.com       en          tech    1            US
#> 218   washingtonpost.com       en          tech None            US
#> 325         usatoday.com       en          tech None            US
#> 500         theverge.com       en          tech    1            US
#> 553       thenextweb.com       en          tech None            US
#> 592          thehill.com       en          tech None            US
#> 719        techradar.com       en          tech    1            US
#> 1190         reuters.com       en          tech None            US
#> 1229          reddit.com       en          tech None            US
#> 1599           nymag.com       en          tech None            US
#> 1739          newser.com       en          tech None            US
#> 1747    newscientist.com       en          tech None            US
#> 1924             mnn.com       en          tech None            US
#> 2021        mashable.com       en          tech    1            US
#> 2405           kinja.com       en          tech None            US
#> 2878          google.com       en          tech None            US
#> 3028         fortune.com       en          tech None            US
#> 3234        engadget.com       en          tech    1            US
#> 3437            digg.com       en          tech    1            US
#> 3478      denverpost.com       en          tech None            US
#> 3548        dailydot.com       en          tech    1            US
#> 3677             cnn.com       en          tech None            US
#> 3686            cnbc.com       en          tech None            US
#> 3796         cbsnews.com       en          tech None            US
#> 3851        buzzfeed.com       en          tech None            US
#> 3876 businessinsider.com       en          tech None            US
#> 4399       seattlepi.com       en          tech None            US
#>                                                                               rss_url
#> 33                                                   https://news.ycombinator.com/rss
#> 71                                                     http://blogs.wsj.com/cio/feed/
#> 115                                                http://feeds.wired.com/wired/index
#> 218                           http://feeds.washingtonpost.com/rss/business/technology
#> 325         http://content.usatoday.com/feed/stories/technologylive%5Ecommunities/rss
#> 500                                              http://www.theverge.com/rss/full.xml
#> 553                                          http://thenextweb.com/section/tech/feed/
#> 592                                          http://thehill.com/taxonomy/term/27/feed
#> 719                                                      http://www.techradar.com/rss
#> 1190                                  http://feeds.reuters.com/reuters/technologyNews
#> 1229                                          http://www.reddit.com/r/technology/.rss
#> 1599                                                    http://nymag.com/rss/Tech.xml
#> 1739                                          http://rss.newser.com/rss/section/7.rss
#> 1747                                               http://feeds.newscientist.com/tech
#> 1924                                           http://www.mnn.com/rss/green-tech-news
#> 2021                                               http://feeds.mashable.com/Mashable
#> 2405                                                       http://front.kinja.com/rss
#> 2878                                https://developers.google.com/web/updates/rss.xml
#> 3028                                                    http://fortune.com/tech/feed/
#> 3234                                             http://www.engadget.com/rss-full.xml
#> 3437                                                    http://digg.com/rss/index.xml
#> 3478                               http://feeds.denverpost.com/dp-business-technology
#> 3548                                     http://www.dailydot.com/feed/summary/latest/
#> 3677                                              http://rss.cnn.com/rss/cnn_tech.rss
#> 3686                              http://www.cnbc.com/id/19854910/device/rss/rss.html
#> 3796                                          http://feeds.cbsnews.com/CBSNewsSciTech
#> 3851                                                 http://www.buzzfeed.com/tech.xml
#> 3876                                    http://www.businessinsider.com/sai/mobile.rss
#> 4399 http://www.seattlepi.com/business/tech/collectionRss/Tech-CP-Headlines-12589.php
#>      GlobalRank
#> 33         1157
#> 71           83
#> 115         146
#> 218          86
#> 325         135
#> 500         225
#> 553         596
#> 592         666
#> 719         715
#> 1190         81
#> 1229         43
#> 1599        662
#> 1739       5629
#> 1747        526
#> 1924       1991
#> 2021        232
#> 2405       1433
#> 2878          2
#> 3028        315
#> 3234        282
#> 3437        183
#> 3478       1016
#> 3548       1971
#> 3677         55
#> 3686        159
#> 3796        212
#> 3851        249
#> 3876        115
#> 4399       2124