Get headlines A helper function to get just the headlines of the feed

Get headlines A helper function to get just the headlines of the feed

get_headlines(
  website = "ycombinator.com",
  topic = NULL,
  rss_table = package_rss
)

Arguments

website
a url of a new source in the format "news.ycombinator.com"

topic
the topic of the feed, by default it is NULL which means it #will fetch the "main" feed. topics are 'tech', 'news', 'business', 'science', #'finance', 'food', 'politics', 'economics', 'travel', 'entertainment', #'music', 'sport', 'world', but not all site have all topics. #use `describe_url("website")` to check for available feeds.

rss_table
a dataframe with urls and rss feeds in case you #need to construct your own out of websites not in the included database. #Be sure to have the same format as the included data. See `R/package_rss.R` #for details.

Value

a dataframe containing the headlines contained in the feed

Examples

Sys.sleep(3) # adding a small time delay to avoid
# simultaneous posts to the API
get_headlines(website = "ycombinator.com", rss_table = package_rss)

#> GET request successful. Parsing...

#> Warning: Predicate functions must be wrapped in `where()`.
#> 
#>   # Bad
#>   data %>% select(is.character)
#> 
#>   # Good
#>   data %>% select(where(is.character))
#> 
#> ℹ Please update your code.
#> This message is displayed once per session.

#>                                                       feed_entries$item_title
#> 1                              The Future of Online Identity Is Decentralized
#> 2                 Rust for JavaScript Developers – Pattern Matching and Enums
#> 3                           SAS Rescue 3 Trapped British Diplomats in Albania
#> 4                                   Do not remain nameless to yourself (1966)
#> 5                              Candy Land Was Invented for Polio Wards (2019)
#> 6                    Cockroachdb/copyist: Mocking an SQL database in Go tests
#> 7                          Building a self-updating profile README for GitHub
#> 8                         Digital Audio on VHS – The Technics SV-P100 [video]
#> 9                                The company that sells lost airplane luggage
#> 10                     US citizens warned they face arbitrary arrest in China
#> 11           Sandboxie-Plus – a fork of Sandboxie with improved functionality
#> 12              A gentle introduction to HDBSCAN and density-based clustering
#> 13                                               MariaDB Temporal Data Tables
#> 14 Lava tubes may be havens for ancient alien life and future human explorers
#> 15                                                               Tour of Rust
#> 16                                                 Ruby lib/irb/easter-egg.rb
#> 17                                                      Dealing with Insomnia
#> 18                                                         Just Too Efficient
#> 19                              Thinking about a New Type System for Perl/Cor
#> 20                Moving from Siloed Organizations to Distributed Individuals
#> 21                                          Linux kernel in-tree Rust support
#> 22                              The growing short case on Facebook and Google
#> 23                Hosting your entire web application using S3 and CloudFront
#> 24                             The illusion of control, and how to give it up
#> 25                   An F-22 test pilot on the Raptor's flight control system
#> 26                                                      The Polymath Playbook
#> 27                                                   How to Understand Things
#> 28               PG: The biggest source of stress for me at YC was running HN
#> 29             Build a No-Slot MIDI Interface on the Apple ][ Game I/O Socket
#> 30                                             Epigrams on Programming (1982)