Get news

Get news

get_news(website = "ycombinator.com", rss_table = package_rss, topic = NULL)

Arguments

website
a url of a new source in the format "news.ycombinator.com"

rss_table
a dataframe with urls and rss feeds in case you #need to construct your own out of webisites not in the included database.

topic
the topic of the feed, by default it is NULL which means it #will fetch the "main" feed. topics are 'tech', 'news', 'business', 'science', #'finance', 'food', 'politics', 'economics', 'travel', 'entertainment', #'music', 'sport', 'world', but not all site have all topics. #use `topics("webiste")` to check for available feeds.

Value

a tibble containing the contents of the rss feed

Examples

Sys.sleep(3) # adding a small time delay to avoid
# simultaneous posts to the API
get_news(website = "ycombinator.com", rss_table = package_rss)

#> GET request successful. Parsing...

#> # A tibble: 30 x 10
#>    feed_title feed_link feed_description feed_pub_date       item_title
#>    <chr>      <chr>     <chr>            <dttm>              <chr>     
#>  1 Hacker Ne… https://… Links for the i… 2020-06-10 09:03:49 Create di…
#>  2 Hacker Ne… https://… Links for the i… 2020-06-10 09:03:49 SeL4 is v…
#>  3 Hacker Ne… https://… Links for the i… 2020-06-10 09:03:49 Writing a…
#>  4 Hacker Ne… https://… Links for the i… 2020-06-10 09:03:49 Stanford …
#>  5 Hacker Ne… https://… Links for the i… 2020-06-10 09:03:49 Logic.ly …
#>  6 Hacker Ne… https://… Links for the i… 2020-06-10 09:03:49 Teaching …
#>  7 Hacker Ne… https://… Links for the i… 2020-06-10 09:03:49 First U.S…
#>  8 Hacker Ne… https://… Links for the i… 2020-06-10 09:03:49 Welcome t…
#>  9 Hacker Ne… https://… Links for the i… 2020-06-10 09:03:49 No Recent…
#> 10 Hacker Ne… https://… Links for the i… 2020-06-10 09:03:49 A ‘Useles…
#> # … with 20 more rows, and 5 more variables: item_link <chr>,
#> #   item_description <chr>, item_pub_date <dttm>, item_category <list>,
#> #   item_comments <chr>