tld_sources A helper function to explore news sources by country (or other TLD)

tld_sources A helper function to explore news sources by country (or other TLD)

tld_sources(tld = NULL, rss_table = package_rss)

Arguments

tld
a country code top level domain in the format ".de"

rss_table
a dataframe with urls and rss feeds in case you #need to construct your own out of webisites not in the included database.

Value

a tibble containing the contents of the rss feed

Examples

tld_sources(tld = "de", rss_table = package_rss)

#> # A tibble: 207 x 7
#>    clean_url  language topic_unified main  clean_country rss_url      GlobalRank
#>    <chr>      <chr>    <chr>         <chr> <chr>         <chr>        <chr>     
#>  1 zdf.de     de       news          1     DE            https://www… 1953      
#>  2 wz.de      de       news          1     DE            https://www… 30055     
#>  3 wiwo.de    de       tech          None  DE            http://blog… 4444      
#>  4 wiwo.de    de       sport         None  DE            http://blog… 4444      
#>  5 wiwo.de    de       economics     1     DE            http://www.… 4444      
#>  6 winfuture… de       tech          1     None          http://www.… 9035      
#>  7 windowsun… de       tech          1     None          http://wind… 159018    
#>  8 weser-kur… de       sport         None  None          http://www.… 16388     
#>  9 weser-kur… de       news          1     None          http://www.… 16388     
#> 10 werra-run… de       news          1     None          http://www.… 489168    
#> # … with 197 more rows