Get headlines A helper function to get just the headlines of the feed

Get headlines A helper function to get just the headlines of the feed

get_headlines(website = "ycombinator.com", rss_table = package_rss)

Arguments

website
a url of a new source in the format "news.ycombinator.com"

rss_table
a dataframe with urls and rss feeds in case you #need to construct your own out of webisites not in the included database.

Value

a tibble containing the headlines contained in the feed

Examples

Sys.sleep(3) # adding a small time delay to avoid
# simultaneous posts to the API
get_headlines(website = "ycombinator.com", rss_table = package_rss)

#> GET request successful. Parsing...

#> Warning: Predicate functions must be wrapped in `where()`.
#> 
#>   # Bad
#>   data %>% select(is.character)
#> 
#>   # Good
#>   data %>% select(where(is.character))
#> 
#> ℹ Please update your code.
#> This message is displayed once per session.

#>                                                            feed_entries$item_title
#> 1                                            Linux Mint drops Ubuntu Snap packages
#> 2                                                                        CUDA 11.0
#> 3                                          Show HN: Runnaroo – A new search engine
#> 4                                                                Flutter for Linux
#> 5                                                     SUSE to Acquire Rancher Labs
#> 6                                                                    Podcast Stats
#> 7                                      EasyOCR: Ready-to-use OCR with 40 languages
#> 8                  A graphical analysis of women's tops sold on Goodwill's website
#> 9               Polynesian DNA Suggests Epic Voyage to South America 800 Years Ago
#> 10       Tauri – toolchain for building secure native apps that have tiny binaries
#> 11                                                             Concorde ‘B’ (2014)
#> 12                                          Browser Font Rendering Inconsistencies
#> 13  ReadMe (YC W15) is hiring a React engineer for our Slate-based markdown editor
#> 14 Tell HN: Google drops blogspot.in, breaking hundreds of thousands of permalinks
#> 15                   Improving performance with SIMD intrinsics in three use cases
#> 16                                                      Show HN: Watchmespendmoney
#> 17                                                             Software 2.0 (2017)
#> 18                                                      Firefox Picture-in-Picture
#> 19                How Developers Stop Learning: Rise of the Expert Beginner (2012)
#> 20                                   The stable marriage problem and modern dating
#> 21  Launch HN: ElectroNeek (YC W20) – Automatically find and automate routine work
#> 22        Git commit accepts several message flags (-m) to allow multiline commits
#> 23                                             Work on Unimportant Problems (2012)
#> 24                              Foreign data wrappers: PostgreSQL's secret weapon?
#> 25                  Investing in Robotics with YC Founder Trevor Blackwell [video]
#> 26                                             Barebones WebGL in 75 lines of code
#> 27                             You could have invented fractional cascading (2012)
#> 28                 Ariane RISC-V CPU – An open source CPU capable of booting Linux
#> 29                                     Live Reloading for a Production Next.js App
#> 30   VCs Promised to Help Black Founders – My Experience Shows a Different Reality