Get headlines A helper function to get just the headlines of the feed

Get headlines A helper function to get just the headlines of the feed

get_headlines(website = "ycombinator.com", rss_table = package_rss)

Arguments

website
a url of a new source in the format "news.ycombinator.com"

rss_table
a dataframe with urls and rss feeds in case you #need to construct your own out of webisites not in the included database.

Value

a tibble containing the headlines contained in the feed

Examples

Sys.sleep(3) # adding a small time delay to avoid
# simultaneous posts to the API
get_headlines(website = "ycombinator.com", rss_table = package_rss)

#> GET request successful. Parsing...

#> Warning: Predicate functions must be wrapped in `where()`.
#> 
#>   # Bad
#>   data %>% select(is.character)
#> 
#>   # Good
#>   data %>% select(where(is.character))
#> 
#> ℹ Please update your code.
#> This message is displayed once per session.

#>                                                             feed_entries$item_title
#> 1                                                        SeL4 is verified on RISC-V
#> 2                                          Create diagrams using code with Graphviz
#> 3                      Writing a full-text search engine using Bloom filters (2013)
#> 4  Logic.ly – Digital logic simulator for teaching logic gates and digital circuits
#> 5                             Teaching my MIT classes with only free/libre software
#> 6                               Stanford Class on Deep Multi-Task and Meta-Learning
#> 7             First U.S. woman to walk in space dives to deepest point in the ocean
#> 8                                                   No Recent Automation Revolution
#> 9                Welcome to the 21st Century: How to Plan for the Post-Covid Future
#> 10   Scraping Roger Ebert’s reviews and finding his favorite movies on Amazon Prime
#> 11                             A ‘Useless’ Perspective That Transformed Mathematics
#> 12                   Windows Privacy Dashboard: GUI for Windows 10 Privacy Settings
#> 13          A tiny static full-text search engine using Rust and WebAssembly (2019)
#> 14                                                           Test Anything Protocol
#> 15                                Unsupervised translation of programming languages
#> 16   Browse hundreds of jobs at actively hiring YC companies with Work at a Startup
#> 17                                   Way to keep your cool running a Raspberry Pi 4
#> 18                                                                 Whistler (Radio)
#> 19   DeepMind introduces ‘EATS’: adversarial, end-to-end approach to text-to-speech
#> 20                       Snabb: 100 Gbit/s pure software switching using Lua (2019)
#> 21                   Olof Palme murder: Sweden identifies man who killed PM in 1986
#> 22                                                       Quotebacks – Quote the Web
#> 23                    Large-Scale Transactional Data Lake at Uber Using Apache Hudi
#> 24                                                          From chaos to free will
#> 25                                          How Birds Evolved From Dinosaurs (2015)
#> 26                        New hints of volcanism under the heart of northern Europe
#> 27                                                 Haiku R1/beta2 has been released
#> 28          Show HN: XgeneCloud – Instant REST and GraphQL APIs on Any SQL Database
#> 29                           Let Game Theory Tell You When It’s Time to Go Shopping
#> 30                             New inline assembly syntax available in Rust nightly