Get headlines A helper function to get just the headlines of the feed

Get headlines A helper function to get just the headlines of the feed

get_headlines(
  website = "ycombinator.com",
  topic = NULL,
  rss_table = package_rss
)

Arguments

website
a url of a new source in the format "news.ycombinator.com"

topic
the topic of the feed, by default it is NULL which means it #will fetch the "main" feed. topics are 'tech', 'news', 'business', 'science', #'finance', 'food', 'politics', 'economics', 'travel', 'entertainment', #'music', 'sport', 'world', but not all site have all topics. #use `describe_url("website")` to check for available feeds.

rss_table
a dataframe with urls and rss feeds in case you #need to construct your own out of websites not in the included database. #Be sure to have the same format as the included data. See `R/package_rss.R` #for details.

Value

a tibble containing the headlines contained in the feed

Examples

Sys.sleep(3) # adding a small time delay to avoid
# simultaneous posts to the API
get_headlines(website = "ycombinator.com", rss_table = package_rss)

#> GET request successful. Parsing...

#> Warning: Predicate functions must be wrapped in `where()`.
#> 
#>   # Bad
#>   data %>% select(is.character)
#> 
#>   # Good
#>   data %>% select(where(is.character))
#> 
#> ℹ Please update your code.
#> This message is displayed once per session.

#>                                                           feed_entries$item_title
#> 1  Show HN: Primo – all-in-one IDE, CMS, component library, static site generator
#> 2                                                  Google to Invest $10B in India
#> 3                          You shouldn't use ENV variables for secret data (2017)
#> 4                             Libtorrent adds support for the WebTorrent protocol
#> 5                                   The Most Remarkable Legacy System I Have Seen
#> 6                             Show HN: A thread hierarchy management library in C
#> 7          Website allows you to experience what it is like to live with dyslexia
#> 8            Your DS18B20 temperature sensor is likely a fake, counterfeit, clone
#> 9                                                       Dark Web Price Index 2020
#> 10                       WindowSwap: View someone's window somewhere in the world
#> 11              Google Search Upgrades Make It Harder for Websites to Win Traffic
#> 12               Uncleftish Beholding: An Uploosening of English Cleanness (2018)
#> 13                 Pulsar – an open-source distributed pub-sub messaging platform
#> 14                           Insecam: Directory of Unsecured Surveillance Cameras
#> 15                                  Writing a software book and making over $100k
#> 16                            Two Black Holes Colliding Not Enough? Make It Three
#> 17                               Self Host a Wiki or Knowledge Base for Your Team
#> 18                              13% of my website visitors block Google Analytics
#> 19                                  11,000 year old ocher mine in underwater cave
#> 20                                                       Text-Only Social Network
#> 21                                     LibreOffice 2020/2025 Marketing Plan [pdf]
#> 22                                          GNU: A Heuristic for Bad Cryptography
#> 23                       Frank Gehry started off building cities with his grandma
#> 24  Exploring how different framings of the same learning task affect performance
#> 25                 Operators and Sidecars Are the New Model for Software Delivery
#> 26                                 Rust Is Surprisingly Good as a Server Language
#> 27                        What Can Bonobos Teach Us About the Nature of Language?
#> 28                                           Apple Silicon: The Passing of Wintel
#> 29                                             Do We Need a Theory of Everything?
#> 30                                      Do not remain nameless to yourself (1966)