Get headlines A helper function to get just the headlines of the feed

Get headlines A helper function to get just the headlines of the feed

get_headlines(
  website = "ycombinator.com",
  topic = NULL,
  rss_table = package_rss
)

Arguments

website
a url of a new source in the format "news.ycombinator.com"

topic
the topic of the feed, by default it is NULL which means it #will fetch the "main" feed. topics are 'tech', 'news', 'business', 'science', #'finance', 'food', 'politics', 'economics', 'travel', 'entertainment', #'music', 'sport', 'world', but not all site have all topics. #use `describe_url("website")` to check for available feeds.

rss_table
a dataframe with urls and rss feeds in case you #need to construct your own out of websites not in the included database. #Be sure to have the same format as the included data. See `R/package_rss.R` #for details.

Value

a tibble containing the headlines contained in the feed

Examples

Sys.sleep(3) # adding a small time delay to avoid
# simultaneous posts to the API
get_headlines(website = "ycombinator.com", rss_table = package_rss)

#> GET request successful. Parsing...

#> Warning: Problem with `mutate()` input `..1`.
#> ✖ Predicate functions must be wrapped in `where()`.
#> 
#>   # Bad
#>   data %>% select(is.character)
#> 
#>   # Good
#>   data %>% select(where(is.character))
#> 
#> ℹ Please update your code.
#> This message is displayed once per session.
#> ℹ Input `..1` is `across(...)`.

#> Warning: Predicate functions must be wrapped in `where()`.
#> 
#>   # Bad
#>   data %>% select(is.character)
#> 
#>   # Good
#>   data %>% select(where(is.character))
#> 
#> ℹ Please update your code.
#> This message is displayed once per session.

#>                                                             feed_entries$item_title
#> 1                              Germany plans to dim lights at night to save insects
#> 2                                                  Y Combinator Startup Library 2.0
#> 3                   Lab mice have a chill, and that may be messing up study results
#> 4                                    Simdjson: Parsing gigabytes of JSON per second
#> 5                        Launch HN: Datafold (YC S20) – Diff Tool for SQL Databases
#> 6                               Show HN: Auth is now available in Supabase (YC S20)
#> 7        OpenCore bootloader – open-sourced Apple UEFI drivers, enabling Hackintosh
#> 8    Usbkill – anti-forensic tool to halt computer when new USB device is connected
#> 9         Show HN: Create your WebAssembly Calling Card–visual creativity with WASM
#> 10                        Georgia School Reopening Photo Even Worse Than It Appears
#> 11                       Google discontinues the Pixel 4, nine months after release
#> 12                               Feds move to block California’s net neutrality law
#> 13     Scientists rename human genes to stop MS Excel from misreading them as dates
#> 14             NeRF in the Wild: reconstructing 3D scenes from internet photography
#> 15                                                       The Math of Card Shuffling
#> 16                                                           Frances Allen has died
#> 17              Twitter web app now runs ES6+, reducing polyfill bundle size by 83%
#> 18                                                 Adam Shostack on Threat Modeling
#> 19            A new design ship in the Philippines is partly powered by wave energy
#> 20                                        How we scaled Google Meet during Covid-19
#> 21          The Largest Commercial Cylindrical Slide Rule Has a Scale Length of 24m
#> 22 Cluster of 295 Chrome extensions caught hijacking Google and Bing search results
#> 23                                                  An Inside Look at CloudKitchens
#> 24                  Pulse oximeters give biased results for people with darker skin
#> 25         Substack (YC W18) is hiring to build a better business model for writing
#> 26                                                                Simon's Algorithm
#> 27                                     Blackstone to Acquire Ancestry.com for $4.7B
#> 28 MicroBlocks: Scratch and Smalltalk-inspired MCU VM for live embedded programming
#> 29                  Show HN: SameTunes – A Music Compatibility Platform for Spotify
#> 30        Take Gitlab to the command line with GLab, an open-source Gitlab CLI tool