Get headlines A helper function to get just the headlines of the feed

Get headlines A helper function to get just the headlines of the feed

get_headlines(website = "ycombinator.com", rss_table = package_rss)

Arguments

website
a url of a new source in the format "news.ycombinator.com"

rss_table
a dataframe with urls and rss feeds in case you #need to construct your own out of webisites not in the included database.

Value

a tibble containing the headlines contained in the feed

Examples

Sys.sleep(3) # adding a small time delay to avoid
# simultaneous posts to the API
get_headlines(website = "ycombinator.com", rss_table = package_rss)

#> GET request successful. Parsing...

#> Warning: Predicate functions must be wrapped in `where()`.
#> 
#>   # Bad
#>   data %>% select(is.character)
#> 
#>   # Good
#>   data %>% select(where(is.character))
#> 
#> ℹ Please update your code.
#> This message is displayed once per session.

#>                                                          feed_entries$item_title
#> 1                                                          CRDTs: The Hard Parts
#> 2                                                     The Modern-Day Mind-Killer
#> 3                         Testing Firefox More Efficiently with Machine Learning
#> 4                                                       How to Understand Things
#> 5                                   Tracking Pico Balloons Using Ham Radio [pdf]
#> 6            A Doctor Who Championed Hand-Washing And Briefly Saved Lives (2015)
#> 7          Favorite Problems: A Practical Framework for Discovering Your Purpose
#> 8              Scientists Say You Can Cancel the Noise but Keep Your Window Open
#> 9                                                        Beware of Being “Right”
#> 10                               Porting a Wolfenstein-type engine to the MEGA65
#> 11                                             Linux kernel in-tree Rust support
#> 12                       Modes, Medians and Means: A Unifying Perspective (2013)
#> 13                                                          Oura Ring 2 Teardown
#> 14                Scientists warn of potential wave of Covid-linked brain damage
#> 15              Show HN: Form-validation.js – JavaScript form validation library
#> 16                                                            Healthy Self-Doubt
#> 17                                      PureDarwin – Darwin-based open source OS
#> 18                        Comet streaking past Earth, providing spectacular show
#> 19                              2000-year-old Iron Age skeleton discovered in UK
#> 20                                          Learnings from a Year of Being Indie
#> 21                                                          Scachs D’amor (2009)
#> 22  A voice separation model that distinguishes multiple speakers simultaneously
#> 23          Big Tech marshals a right-leaning army of allies for antitrust fight
#> 24 A Little Startup That Could: Aquarius Builds an Engine Others Only Dreamed Of
#> 25                                                        Scraping Isn't Hacking
#> 26 Show HN: Tiny password manager with all data stored encrypted on your machine
#> 27                          Commander X16 8 bit computer now has its own website
#> 28                                           The Frink Is Good, the Unit Is Evil
#> 29                       'Street Fighter II' Wasn’t Just a Game, It Was a Portal
#> 30                         Don't close your MacBook with a cover over the camera